+++ NEWS.md
@@ -4,7 +4,7 @@
 * Improved handling of `get_diff()` response data
 * Improved handling of `page_vector_functions()` response data. Most functions now return a tbl or list of tbls
 * Basic support for `wikimedia_rest_apis()`
-* 'peform_query_once()' renamed to `next_batch()`
+* 'perform_query_once()' renamed to `next_batch()`
 
 # wikkitidy 0.1.1